modules-v1: Implement the download endpoint
This is still not actually compliant with the currently-specified protocol
because we are supposed to return the URL of the archive rather than
directly the archive itself. This service doesn't know its own absolute
URL, so we'll need to change the protocol to support some sort of
relative addressing or direct download before this can be used.