config: require "hostname" to be set for module registries
The registry protocol requires a registry to know its own hostname so
that it can produce canonical module source strings in the "version"
response.

Here we use Terraform's own svchost package to process the given hostname,
which ensures that it'll be interpreted the same way that Terraform
would interpret it.